requirements.txt
langchain
langchain-google-genai
langchain-community
langgraph
tavily-python
python-dotenv
pillow